api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0.0
httpx
